refactor: extract PDF table line band analysis
